+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.9.1 2020/11/30
+
+Break Change
+- [beaconscan][beaconchain] switch from medalla testnet to mainnet
+
+Fix
+- fix command [gas][tx][valiidator]
+
 # 0.8.2 2020/11/28
 
 Feature